+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update Vulkan Headers to version 183
+- Use a `SmallVec` instead of a `Vec` when returning collections
 
 ### Fixed